+++ Project0/src/com/project0/finances/Payments.java
@@ -1,179 +1,58 @@
 package com.project0.finances;
 
-import java.util.HashMap;
-import java.util.Scanner;
-
-import com.project0.utilities.ScannerUse;
 
 public class Payments {
-
-	public Payments() {
-
+	private int accountId;
+	private int carId;
+	private String cusUsername;
+	private double monthlyPayments;
+	private double payOffbalance;
+	public int getAccountId() {
+		return accountId;
+	
+	
+	
+	}
+	public Payments(int accountId, int carId, String cusUsername, double monthlyPayments, double payOffbalance) {
 		super();
-
+		this.accountId = accountId;
+		this.carId = carId;
+		this.cusUsername = cusUsername;
+		this.monthlyPayments = monthlyPayments;
+		this.payOffbalance = payOffbalance;
 	}
-
-	/*
-	 * 
-	 * I created a scanner to implement in all the methods below.
-	 * 
-	 * just check it out and see if that is correct for your scanner class
-	 * 
-	 * to determine if it will work or not.
-	 * 
-	 * im getting a few errors, im not sure why but then hopefully the code runs.
-	 * 
-	 * if you see any changes that need to be done let me know.
-	 * 
-	 */
-	 ScannerUse su = ScannerUse.getInstance();
-	 Scanner cs =su.returnScanner();
-
-
-
-	public HashMap<String, Double> carlist;
-
-	public double principal;
-
-	public int term;
-
-	public double payments;
-
-	public double rate;
-
-	public int carLot() {
-
-		HashMap<String, Double> carList = null ;
-
-		carList.put("Subaru", 29095.00);
-
-		carList.put("Ford", 29965.00);
-
-		carList.put("Cadillac", 63215.00);
-
-		carList.put("BMW", 40499.00);
-
-		carList.put("Jeep", 50317.00);
-
-		carList.put("Chevrolet", 49000.00);
-
-		System.out.println(carList);
-
-		return carList.size();
-
+	public void setAccountId(int accountId) {
+		this.accountId = accountId;
 	}
-
-	public double annualTerms(int term) {
-
-		// number of months the auto loan will be
-
-		// customers will be able to select the terms of their loan
-
-		// need to use the scanner method
-
-		double[] arrterm = { 24, 36, 48, 60, 72, 84 };
-
-		for (int i = 0; i <= arrterm.length; i++) {
-
-			System.out.println(term);
-
-		}
-
-		System.out.println("Select the term that fits your desire.");
-
-		term = cs.nextInt();
-
-		System.out.println("Customer elected to term condition: " + term);
-
-		return term;
-
+	public int getCarId() {
+		return carId;
 	}
-
-	public double carLoan(double principal) {
-
-		HashMap<String, Double> carList = null;
-
-		if (carList.isEmpty()) {
-
-			System.out.println("Car Lot is vacant.");
-			;
-
-		} else if (carList.size() > 0) {
-
-			// need to implement the scanner method so customer can select car.
-
-			System.out.println("Select from Vehicle List: " + carList.entrySet());
-
-			String vehicleselect = cs.nextLine();
-
-			System.out.println("You have selected: " + vehicleselect);
-
-		}
-
-		//String vehicleselect;
-
-		if ("vehicleselect" == "Subaru") {
-
-			System.out.println(carList.get("Subaru"));
-
-		} else if ("vehicleselect" == "Ford")
-
-			System.out.println(carList.get("Ford"));
-
-		else if ("vehicleselect" == "Cadillac") {
-
-			System.out.println(carList.get("Cadillac"));
-
-		} else if ("vehicleselect" == "BMW") {
-
-			System.out.println(carList.get("BMW"));
-
-		} else if ("vehicleselect" == "Jeep") {
-
-			System.out.println(carList.get("Jeep"));
-
-		} else if ("vehicleselect" == "Chevrolet") {
-
-			System.out.println(carList.get("Chevrolet"));
-
-		} else {
-
-			System.out.println("You have to make a selection.");
-
-		}
-
-		return principal;
-
+	public void setCarId(int carId) {
+		this.carId = carId;
 	}
-
-	public double interestRate(double rate) {
-
-		System.out.println("Select interest rate from 1 to 20.");
-
-		double intrate = cs.nextDouble();
-
-		System.out.println("You've selected: " + intrate);
-
-		double a = 12.0;
-
-		rate = (intrate / a);
-
-		return rate;
-
+	public String getCusUsername() {
+		return cusUsername;
 	}
-
-	public double monthlyPayments(double monthlyPayment) {
-
-		double annualrate = this.interestRate(rate);
-
-		double vehicleprice = this.carLoan(principal);
-
-		double months = this.annualTerms(term);
-
-		monthlyPayment = ((annualrate * vehicleprice) / (1 - (1 + annualrate) * (-(Math.sqrt(months)))));
-
-		return monthlyPayment;
-
+	public void setCusUsername(String cusUsername) {
+		this.cusUsername = cusUsername;
+	}
+	public double getMonthlyPayments() {
+		return monthlyPayments;
+	}
+	public void setMonthlyPayments(double monthlyPayments) {
+		this.monthlyPayments = monthlyPayments;
+	}
+	public double getPayOffbalance() {
+		return payOffbalance;
+	}
+	public void setPayOffbalance(double payOffbalance) {
+		this.payOffbalance = payOffbalance;
+	}
+	@Override
+	public String toString() {
+		return "Payments [accountId=" + accountId + ", carId=" + carId + ", cusUsername=" + cusUsername
+				+ ", monthlyPayments=" + monthlyPayments + ", payOffbalance=" + payOffbalance + "]";
 	}
 
 }
+
